+++ config/webpack/dev.js
@@ -39,6 +39,12 @@ var config = {
 
   module: {
     rules: [{
+        test: /i18n\/locales/,
+        loader: '@alienfast/i18next-loader',
+        query:{
+          relativePathAsNamespace: true,
+        }
+      }, {
         enforce: 'pre',
         test: /\.tsx?$/,
         loader: 'tslint-loader'
