+++ src/i18n/crowdin.yml
@@ -1,0 +1,16 @@
+"project_id" : "442032"
+"api_token_env" : "CROWDIN_PERSONAL_TOKEN"
+"base_path" : "."
+"base_url" : "https://api.crowdin.com"
+
+"preserve_hierarchy": true
+
+files: [
+ {
+  "source" : "src/i18n/locales/en/translation.json",
+  "translation" : "src/i18n/locales/%two_letters_code%/translation.json",
+  "dest" : "app.json",
+  "type" : "json",
+  "skip_untranslated_strings": true,
+ }
+]
