+++ src/i18n/locales/index.js
@@ -1,0 +1,2 @@
+// hack needed for @alienfast/i18next-loader
+module.exports = {};
