+++ src/i18n/setup.ts
@@ -1,30 +1,28 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
+import Pseudo from 'i18next-pseudo';
+import resources from './locales';
 
 export function setupI18n() {
+  console.log(resources);
   i18n
     .use(initReactI18next)
+    .use(new Pseudo({
+      enabled: true,
+      languageToPseudo: 'tlh',
+    }))
     .use(LanguageDetector)
     .init({
       detection: {
         order: ['querystring', 'localStorage', 'navigator'],
         caches: [],
       },
-      resources: {
-        en: {
-          translation: require('./en.json'),
-        },
-        cy: {
-          translation: require('./cy.json'),
-        },
-        pt: {
-          translation: require('./pt.json'),
-        },
-      },
+      resources,
       fallbackLng: 'en',
       interpolation: {
         escapeValue: false,
       },
+      postProcess: ['pseudo'],
     });
 }
