--- src/locales/cy.json
@@ -1,7 +1,0 @@
-{
-  "Beneficiary": "Beneficiary",
-  "Log Out": "Log Out",
-  "Questionnaires": "Questionnaires",
-  "Report": "Report",
-  "Settings": "Settings"
-}
